Moved ElasticSearchClient into package, added docstrings, skip ordinary unit tests when pytest is run with --runelasticsearch
